requirements.txt
langchain>=0.1.14
openai>=1.3.5
tiktoken>=0.5.1
pydantic>=2.6
python-dotenv>=1.0
toml>=0.10
psycopg[binary]>=3.1
pytest>=8.0
